+++ CHANGELOG.md
@@ -7,6 +7,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [0.0.30] - 2023-08-23
+
 ### Changed
 - Extend proposal grpc response with timeline field